IOS-817 if something goes wrong on login the account is removed

+++ pEpForiOS/UI/AccountsAndFolders/AccountsViewModel/LoginViewModel.swift
@@ -133,9 +133,9 @@ extension LoginViewModel: AccountVerificationServiceDelegate {
     func verified(account: Account, service: AccountVerificationServiceProtocol,
                   result: AccountVerificationResult) {
         mySelfer?.startMySelf()
-        /*if result != .ok {
+        if result != .ok {
             account.delete()
-        }*/
+        }
         delegate?.didVerify(result: result)
     }
 }
